Store all variable values by default
Previously the default configuration was to not store the full timeseries
for time indexed variables but instead only store the current and previous
values so that memory usage is low. This is changed so that by default all
values are stored and one has to manually switch to the efficient storage
implementation if one needs it.